Refactor resource_manager::stop()
The semantics of the function has been changed.
If space_watchdog::stop() throws an exception, that will be the only exception being thrown.
In the case of manager::stop() throwing an exception, the original implementation would have
thrown a nested exeption if space_watchdog::stop() also fails. However, it does not look
like space_watchdog::stop() could ever throw an exception, even when co_awaited,
so I went with this change.