Get rid of a possible dangerous use-after-free
Commitlog converts functors passed to it to std::function. This has the benefit
of controlling the lifetime of those captures that were captured by value.
However, I doubt many developers pay attention to the signature of the function.
If it changes, it can create various problems, one of which is related to passing
a coroutine lambda. The person changing the interface in the future would, of course,
be responsible for making sure other code does not break, but we can ensure that on our own.
Using seastar::coroutine::lambda, whose semantics is similar to passing a lambda object by
std::ref, also triggers the small object optimization in std::function.