db/hints: Move send_one_file_ctx to hint_sender.cc
The struct is only used internally by hint_sender.
Though it appears in the API as an argument of some functions,
it is wrapped in seastar::lw_shared_ptr, which accepts
incomplete types and those functions are private.
Because of that, there is no reason to keep it in the header.