db/hints: Refactor wait_for_sync_point()
This commit changes the range we iterate through.
Instead of iterating over endpoint managers, we start
iterating over shard_rps. The rationale for that is
that usually, the size of shard_rps should be smaller
than the number of endpoint managers currently
managed by the hint shard manager. Also, if no endpoint
manager is found, we immediately return.