db/hints/manager: Divide start method
This commit is part of the commit series whose goal
is to postpone starting the hint managers so that
they have access to a valid topology instance
and all endpoint managers start properly.

In this particular commit, we divide the start
method to draw a line between setting up
the hint manager instances and starting them.