db/hints/resource_manager: Divide start method
This commit is part of the commit series whose goal
is to postpone starting the hint managers so that
they have access to a valid topology instance
and all endpoint managers start properly.

In this commit, we split the start method of
resource manager into setting it up (assigning
the gossiper) and starting the hint managers.