main: Postpone start-up of hint manager
In this commit, we postpone the start-up
of the hint manager until we obtain information
about other nodes in the cluster.

The rationale for these changes is to make sure
that db::hints::manager::start() has access
to a locator::topology as complete as possible.
It's crucial because based on its state, we may
or may not start db::hints::internal::endpoint_manager's.
That might prevent us from sending hints even
if the corresponding node is up until we decide
to store at least one hint for that node -- only
then would the endpoint manager be started.

These changes aim to solve that issue.