service: Add locator::host_id to on_leave_cluster
We extend the function
endpoint_lifecycle_subscriber::on_leave_cluster
by another argument -- locator::host_id.
The rationale for these changes is that
we need to know the host ID of a node
that is leaving the cluster, but it
might not be part of locator::token_metadata
at that point anymore. This commit solves
that issue.