service: Add overload of on_leave_cluster for locator::host_id
This commit is part of the series migrating
Hinted Handoff from using IPs to using
locator::host_id.

We add an overload of the function
endpoint_lifecycle_subscriber::on_leave_cluster
to accept locator::host_id as its argument.
The rationale for these changes is that
we need to know the host ID of a node
that is leaving the cluster, but it
might not be part of locator::token_metadata
at that point anymore. To solve the problem,
we extend the API to be able to pass it
explcitily.