db/hints: Migrate directories to host ID during upgrade
During start-up, we rename existing directories
from IPs to host IDs. If renaming them is
impossible, we remove such a directory.
While that may seem suspicious and error-prone,
Hinted Handoff is a best-effort mechanism.
We will only need to perform the mapping
of the directories' names once, so though
it's feasible to handle it 'better', it
would make the start-up process more complex.
We decide not to do that.