db/hints: Prepare Hinted Handoff to use IP and host ID
This commit introduces changes necessary for
creating and managing hint directories
when they're identified by either IPs or
host IDs.

We start keeping track of mappings between
IPs and host IDs where the IPs correspond
to hint directories and host IDs correspond
to endpoint managers managing them.

When a hint to a target with no endpoint
manager comes to the hint manager, we
save a mapping of the associated pair
of the hint's target's IP and host ID
in a map. This way, we always have
a way to associate a managed hint
directory's name with its endpoint
manager.