db/hints: Set state::draining_all early on
We mark a shard hint manager with draining_all
early on in manager::drain_for() early on,
before the coroutine can yield, if the argument
corresponds to the local node. The rationale is
that we need that information -- we can't start
the migration process while the local node is
being drained, but it shouldn't happen afterwards
either as the shard manager won't accept any
new hints from that point on. The node is
effectively stopped.