db/hints: Keep track of drained nodes
We start storing mappings host ID -> future<>
corresponding to drained nodes. We can't
stop the same endpoint manager twice because
an exception will be issued, but during
the migration process we need to make sure
all endpoint managers have been stopped.
This is a solution to that problem
-- if draining of a node has started,
we won't stop it, and we will await the
correspodning future instead.