cql3/statements/describe_statement: Add new functions to generate descriptions
We add new functions to generalize existing `generate_decriptions()` for
types that inherit from the abstract class `data_dictionary::describable_entity`
or the abstract class `data_dictionary::description_generator` -- and also
ranges of those.

The rationale for that is that the existing `generator_description()` requires
that the types passed to it inherit from `data_dictionary::keyspace_element`.
However, not all entities that can be described belong to any keyspace.
For example, in the following commits we will start describing roles, role
grants, service levels, etc.