cql3/describe_statement: Extend DESC SCHEMA with information on AUTH and service levels
When executing `DESC SCHEMA WITH INTERNALS`, Scylla now also returns
statements that can be used to restore the state of auth
and recreate service levels. That encompasses granting roles
and permissions as well as attaching service levels to roles.

If the additional parameter `WITH SALTED HASHES` is provided,
the statements corresponding to recreating roles in the system
will also contain the stored salted hashes.