auth/authenticator: Add member functions for querying salted hash
We add new member functions to the interface of `auth::authenticator`
responsible for querying the salted hash corresponding to a given
role. One method indicates whether a given authenticator uses
salted hashes, while the other queries them or throws an exception
salted hashes are not used.

The rationale for extending the interface of authenticator is
to be able to access salted hashes from other parts of auth.
We will need them in an upcoming commit responsible for describing
auth.