auth: Reject invalid hashed passwords
Before these changes, the user could create a role with an almost
arbitrary hashed password. This patch implements an additional check
that ensures that the provided hashed password matches the format of one
of the encryption algorithms supported by Scylla: bcrypt, sha512,
sha256, or md5. In case it doesn't, Scylla throws an exception.

We add tests verifying the change is correct and modify the existing
ones to adjust them to the current semantics of the statement.