test/pylib/repair.py: Assign nodes to multiple racks in create_table_insert_data_for_repair
We assign the newly created nodes to multiple racks. If RF <= 3,
we create as many racks as the provided RF. If RF > 3, exactly
three racks are created. The rationale for doing this is we want
to ensure that the keyspace we're going to create next will be
RF-rack-valid and so Scylla will allow for it even with the
`rf_rack_valid_keyspaces` configuration option enabled.