mv, index: Allow for creating MVs and indexes only in RF-rack-valid keyspaces
We allow for creating materialized views and secondary indexes only in
RF-rack-valid keyspaces. That extends to start-up: Scylla will refuse
to start if there is an MV/index using tablets, but it's not started with
the `rf_rack_valid_keyspaces` configuration option enabled.

This commit also provides tests verifying the implementation.