mv, index: Replace experimental feature with rf_rack_valid_keyspaces
Before these changes, creating materialized views and secondary indexes
was only possible when Scylla was started using an experimental feature
-- `views-with-tablets`.

In this commit, we get rid of the feature, mark it as unused, and replace
the restriction with the `rf_rack_valid_keyspaces` configuration option:
creating a materialized view or a secondary index in keyspaces using tablets
will be now impossible unless `rf_rack_valid_keyspaces` is enabled.

We provide a test verifying the change.