test/boost/tablets_test.cc: Adjust test_load_balancing_works_with_in_progress_transitions to RF-rack-validity
We distribute the nodes used in the test across two racks. That may
affect how tablets behave and it does. That's why it's important
that we assign to nodes, both with 1 shard per node, to one rack,
while the remaining one node to a separate rack. Thanks to that,
the load balancer will allow for migrations to proceed as expected.