cql3/description: Introduce deserialize_create_statement()
We introduce a new method and start using it in places where
`create_statement` is not used for serialization purposes.
For now, the change is almost a NO-OP, but we're going to
change the underlying type of `cql3::description::create_statement`
and this commit is supposed to make the transition simpler.