bytes_ostream.hh: Overload write() and introduce operator<<
We extend the interface of `bytes_ostream` to use it in the
following commit. That will make the changes much simpler.

This change may be controversial since it allows for writing
regular `char`s to a stream of bytes; the types are isomorphic
in most cases, but they're not the same. However, the existing
API of the class already allows for that: we have the overload
`write(const char*, size_t)`. The change this commit brings
is yet another form of it.