cql3/functions/user_function.cc: Use std::ostringstream when describing
We modify the code responsible for describing a UDF so that it uses
std::ostringstream instead of std::ranges. It may seem strange as we're
generally opting for using ranges in the codebase more and more as it
often provides a clearer overview of what's going on. However, this change
is a preliminary for replacing the stream object with one that internally
uses a fragmented buffer -- we'll implement it in an upcoming commit.

Refs scylladb/scylladb#24018