db/schema_tables.cc: Remove code covering altering secondary index
The auxiliary function `make_update_indices_mutations` is responsible for
tracking changes in the secondary indexes when performing a schema change
and reacting to them. It covers all possible cases:

(1) a new index appears,
(2) an existing index disappears,
(3) or an existing index changes.

To that end, the function compares two sets of indexes in the form of
mappings `index_name -> index_metadata`: before a schema change and after it.

However, at the moment, there is no way to lead to case (3):

* The only allowed CQL operations on an index are CREATE and DROP. There's no
  ALTER INDEX statement. In other words, we can only lead to case (1) or (2).
* Changes in the underlying materialized view are mostly ignorant of the index
  that uses it. We won't be able to drop it in any other way than by performing
  DROP INDEX, but modifying its options will have no effect on the index metadata.
* Similarly, changes in the base table should not lead to any modification of
  the index metadata of any index based on it. Renaming columns is an interesting
  case, though:
  * we can only rename part of the primary key, but we cannot create an index
    on the primary key,
  * the clustering key is NOT eligible for renaming,
  * if the primary key consists of multiple columns, e.g. when creating a table like
    this:
    ```
    CREATE TABLE ks.t (p int, v int, u int, PRIMARY KEY((p, v)));
    ```
    we should be able to rename `v`, but if there's a secondary index on that column,
    it won't be possible:
    ```
    InvalidRequest: Error from server: code=2200 [Invalid query] message="Cannot rename
                    column v because it has dependent secondary indexes (i)"
    ```
* Modifying index metadata via a mutation to `system_schema.indexes` is a regular
  mutation, not	a schema change, so it won't lead to a difference in two instances
  of metadata in `make_update_indices_mutations`. The user shouldn't attempt to modify
  it directly anyway.

Because of that, case (3) can be dropped in the code. There's no harm in keeping it,
but we're going to refactor the logic responsible for managing indexes in upcoming
commits and this commit is a stepping stone.